+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Clearing an ability score turns it to a null (-) value
 - Level 0 spells now use up charges, but will be migrated to not auto deduct charges this version
 - Adjusted when players can see a token's name in the tooltip slightly, so that they can always see the actual name if they at least have Observer permissions
+- Added `/damage <roll>` and `/heal <roll>` chat commands
 
 ## 0.77.9 (Hotfix)